requirements.txt
click>=8.0.0
jmespath>=1.0.0
pyyaml>=6.0
rich>=10.0.0
python-hcl2>=3.0.0
jsonschema>=4.0.0
pyinstaller>=5.0
